fix ull in popLsb

+++ src/Board.cpp
@@ -1205,7 +1205,7 @@ public:
     }
 
 
-    int popLsb(int &bitBoard) {
+    int popLsb(ull &bitBoard) {
         //get LS 1 in the board and toggle itpop
         ull z  = (log2(bitBoard & -bitBoard) + EPS);
         bitBoard = ((bitBoard & -bitBoard) ^ bitBoard);
